+++ extensions/api/data-management/asset/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/asset/service/AssetService.java
@@ -35,7 +35,7 @@ public interface AssetService {
      * Query assets
      *
      * @param query request
-     * @return the collection of assets that match the query
+     * @return the collection of assets that matches the query
      */
     Collection<Asset> query(QuerySpec query);