+++ extensions/api/data-management/asset/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/asset/service/AssetServiceImpl.java
@@ -84,7 +84,7 @@ public class AssetServiceImpl implements AssetService {
         var result = new AtomicReference<ServiceResult<Asset>>();
 
         transactionContext.execute(() -> {
-            if (index.findById(asset.getId()) == null) {
+            if (findbyId(asset.getId()) == null) {
                 loader.accept(asset, dataAddress);
                 result.set(ServiceResult.success(asset));
             } else {