+++ CHANGELOG.md
@@ -29,11 +29,12 @@ the detailed section referring to by linking pull requests or issues.
 * Implement S3BucketReader (#675)
 * Add configuration setting for state machine batch size (#872)
 * Add Jetty context alias for IDS API (#815)
-* Implement Asset service for Data Management API (#931)
-* Added embedded and remote DPF Selector (#832)
+* Add embedded and remote DPF Selector (#832)
 * Pass path information into http data source through DPF public API (#929)
 * Add instructions for observability sample with Azure Application Insights (#928)
 * Add interface `WebServer` to `web-spi` (#921)
+* Implement Asset service for Data Management API (#931)
+* Implement ContractDefinition service for Data Management API (#940)
 
 #### Changed